refactor: 상품 컨트롤러 DTO record로 변경 및 패키지 이동
