+++ src/main/java/kr/hhplus/be/server/interfaces/item/ItemApiSpec.java
@@ -6,8 +6,6 @@ import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import kr.hhplus.be.server.interfaces.item.dto.ItemResponse;
-import kr.hhplus.be.server.interfaces.item.dto.ItemRequest;
 import org.springframework.http.ResponseEntity;
 
 import java.util.List;
